+++ core-web/libs/sdk/client/src/lib/editor/sdk-editor-vtl.ts
@@ -1,12 +1,24 @@
-import { initEditor, isInsideEditor } from './sdk-editor';
+import {
+    listenContentChange,
+    listenEditorMessages,
+    listenHoveredContentlet,
+    pingEditor
+} from './listeners/listeners';
+import { isInsideEditor } from './sdk-editor';
 
 /**
- * This is the main entry point for the SDK VTL. It initializes the client and returns it.
+ * This is the main entry point for the SDK VTL.
  * This is added to VTL Script in the EditPage
  *
- * @type {*}
+ * @remarks
+ * This module sets up the necessary listeners and functionality for the SDK VTL.
+ * It checks if the script is running inside the editor and then initializes the client by pinging the editor,
+ * listening for editor messages, hovered contentlet changes, and content changes.
+ *
  */
-
 if (isInsideEditor()) {
-    initEditor();
+    pingEditor();
+    listenEditorMessages();
+    listenHoveredContentlet();
+    listenContentChange();
 }
