--- dotCMS/src/main/webapp/html/js/editor-js/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@dotcms/client",
-  "version": "0.0.1-alpha.8",
-  "type": "module",
-  "description": "Official JavaScript library for interacting with DotCMS REST APIs.",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/dotCMS/core.git#master"
-  },
-  "keywords": [
-    "dotCMS",
-    "CMS",
-    "Content Management",
-    "API Client",
-    "REST API"
-  ],
-  "author": "dotcms <dev@dotcms.com>",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/dotCMS/core/issues"
-  },
-  "homepage": "https://github.com/dotCMS/core/tree/master/core-web/libs/sdk/client/README.md",
-  "module": "./sdk-editor.esm.js",
-  "main": "./sdk-editor.esm.js"
-}
