--- dotCMS/src/main/webapp/html/js/editor-js/sdk-editor.esm.d.ts
@@ -1,1 +1,0 @@
-export * from "./src/lib/editor/sdk-editor-vtl";
