--- dotCMS/src/main/webapp/html/js/editor-js/src/index.d.ts
@@ -1,4 +1,0 @@
-export * from './lib/client/sdk-js-client';
-export * from './lib/editor/sdk-editor';
-export * from './lib/editor/models/editor.model';
-export * from './lib/editor/models/client.model';
