--- dotCMS/src/main/webapp/html/js/editor-js/src/lib/editor/sdk-editor-vtl.d.ts
@@ -1,1 +1,0 @@
-export {};
