--- dotCMS/src/main/webapp/html/js/editor-js/src/lib/editor/sdk-editor.d.ts
@@ -1,23 +1,0 @@
-import { DotCMSPageEditorConfig } from './models/editor.model';
-/**
- *
- * Updates the navigation in the editor.
- * @param {string} pathname - The pathname to update the navigation with.
- * @memberof DotCMSPageEditor
- */
-export declare function updateNavigation(pathname: string): void;
-/**
- * Checks if the code is running inside an editor.
- * @returns {boolean} Returns true if the code is running inside an editor, otherwise false.
- */
-export declare function isInsideEditor(): boolean | undefined;
-/**
- * Initializes the DotCMS page editor.
- *
- * @param conf - Optional configuration for the editor.
- */
-export declare function initEditor(config?: DotCMSPageEditorConfig): void;
-/**
- * Destroys the editor by removing event listeners and disconnecting observers.
- */
-export declare function destroyEditor(): void;
