+++ ema/nextjs/src/app/[[...slug]]/page.js
@@ -38,6 +38,6 @@ export default async function Home({ searchParams, params }) {
         depth: 2,
         languageId: searchParams.language_id
     });
-    
+
     return <MyPage nav={nav.entity.children} data={data.entity}></MyPage>;
 }
